% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdec.R
\name{gls.beta.sig2}
\alias{gls.beta.sig2}
\title{C++ interface for GLS}
\usage{
gls.beta.sig2(x, y, id, S, omega)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{vector of responses}

\item{id}{vector of cluster discriminators}

\item{S}{vector of coordinates within clusters (used for intra-cluster distances)}

\item{omega}{2-vector of correlation model parameters}
}
\description{
C++ interface for GLS
}
