% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdec.R
\name{print.summary.rdec}
\alias{print.summary.rdec}
\title{print for summary}
\usage{
\method{print}{summary.rdec}(x, digits = max(3, .Options$digits - 3), ...)
}
\arguments{
\item{x}{instance of summary.rdec}

\item{digits}{numeric(2)}

\item{\dots}{not used}
}
\description{
print for summary
}
