% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdec.R
\name{quadform.bd.dec}
\alias{quadform.bd.dec}
\title{C++ interface for banded matrix}
\usage{
quadform.bd.dec(x, id, S, omega)
}
\arguments{
\item{x}{matrix of predictors}

\item{id}{vector of cluster discriminators}

\item{S}{vector of coordinates within clusters (used for intra-cluster distances)}

\item{omega}{2-vector of correlation model parameters}
}
\description{
C++ interface for banded matrix
}
