% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdec.R
\name{rdec.m2llik}
\alias{rdec.m2llik}
\title{likelihood evaluator}
\usage{
rdec.m2llik(omega, x, y, id, S, beta, sig2)
}
\arguments{
\item{omega}{2-vector of (gamma,theta)}

\item{x}{matrix of predictors}

\item{y}{vector of responses}

\item{id}{vector of cluster discriminators}

\item{S}{vector of coordinates within clusters (used for intra-cluster distances)}

\item{beta}{regression parameter vector}

\item{sig2}{variance estimate}
}
\description{
likelihood evaluator
}
