% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdec.R
\name{summary.rdec}
\alias{summary.rdec}
\title{summary report}
\usage{
\method{summary}{rdec}(object, ...)
}
\arguments{
\item{object}{an rdec S3 instance}

\item{\dots}{can include wt, a vector of weights}
}
\description{
summary report
}
