% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdec.R
\name{rdec}
\alias{rdec}
\title{regression with damped exponential correlation}
\usage{
rdec(
  formula,
  id,
  S,
  data,
  subset,
  na.action = na.fail,
  omega.init = c(0, 0),
  omega.low = c(0.001, 0),
  omega.high = c(0.95, 0.95),
  ltol = 0.01,
  contrasts = NULL,
  verbose = FALSE
)
}
\arguments{
\item{formula}{standard linear modeling formula}

\item{id}{vector of cluster discriminators}

\item{S}{vector of observation times}

\item{data}{resolve model formula elements in this data frame}

\item{subset}{logical subsetting vector}

\item{na.action}{action to take for NA in formula element}

\item{omega.init}{starting value for covariance parameter vector (dim=2)}

\item{omega.low}{lower bound for covariance parameter vector}

\item{omega.high}{upper bound for covariance parameter vector}

\item{ltol}{tolerance for convergence of likelihood function}

\item{contrasts}{probably not used}

\item{verbose}{print -2 likelihood value per iteration}
}
\value{
a structure with model fitting results
}
\description{
ML fitting of a multivariate gaussian model for longitudinal data that
generalizes AR(1), CS, and MA(1) covariance structures
}
\note{
uses optim with specified gradient function
}
\examples{

data(rdtest)
summary( rdec(rdec.y~rdec.x, id=rdtest$rdec.id, S=rdtest$rdec.s, omega.init = c(.5,.5), omega.low=c(.01,.01),
   omega.high=c(.95, .95), data=rdtest ))

}
\references{
Munoz Carey Rosner et al, Biometrics 1992
}
\seealso{
gee, geepack, nlme, reshape
}
\author{
VJ Carey <stvjc@channing.harvard.edu>
}
\keyword{models}
